+++ R/summary.R
@@ -16,6 +16,9 @@ print.RoBMA <- function(x, ...){
 }
 
 
+
+
+
 #' @title Summarize fitted RoBMA object
 #'
 #' @description \code{summary.RoBMA} creates summary tables for a
@@ -40,6 +43,9 @@ print.RoBMA <- function(x, ...){
 #' (couple) of letters. Defaults to \code{FALSE}.
 #' @param remove_spike_0 whether spike prior distributions with location at zero should
 #' be omitted from the summary. Defaults to \code{FALSE}.
+#' @param standardized_coefficients whether to show standardized meta-regression coefficients.
+#' Defaults to \code{TRUE}. When set to \code{FALSE}, raw (unstandardized) meta-regression
+#' coefficients are returned for continuous predictors. Only affects meta-regression models.
 #' @param ... additional arguments
 #'
 #' @examples \dontrun{
@@ -63,6 +69,11 @@ print.RoBMA <- function(x, ...){
 #'
 #' # summary of individual models and their parameters can be further obtained by
 #' summary(fit, type = "individual")
+#'
+#' # for meta-regression models, raw (unstandardized) coefficients can be obtained with
+#' # data(Kroupova2021)  # example data with predictors
+#' # fit_reg <- RoBMA.reg(r ~ age, data = Kroupova2021[1:20,])
+#' # summary(fit_reg, standardized_coefficients = FALSE)
 #' }
 #'
 #' @note See [diagnostics()] for visual convergence checks of the individual models.
@@ -74,7 +85,7 @@ print.RoBMA <- function(x, ...){
 #' @export
 summary.RoBMA       <- function(object, type = "ensemble", conditional = FALSE,
                                 output_scale = NULL, probs = c(.025, .975), logBF = FALSE, BF01 = FALSE,
-                                short_name = FALSE, remove_spike_0 = FALSE, ...){
+                                short_name = FALSE, remove_spike_0 = FALSE, standardized_coefficients = TRUE, ...){
 
   # apply version changes to RoBMA object
   object <- .update_object(object)
@@ -87,6 +98,7 @@ summary.RoBMA       <- function(object, type = "ensemble", conditional = FALSE,
   BayesTools::check_bool(logBF, "logBF")
   BayesTools::check_bool(short_name, "short_name")
   BayesTools::check_bool(remove_spike_0, "remove_spike_0")
+  BayesTools::check_bool(standardized_coefficients, "standardized_coefficients")
 
   # check the scales
   if(is.null(output_scale)){
@@ -106,7 +118,7 @@ summary.RoBMA       <- function(object, type = "ensemble", conditional = FALSE,
 
   # dispatch the spike and slab summary function if needed
   if(object[["add_info"]][["algorithm"]] == "ss"){
-    return(.summary.RoBMA.ss(object, type, conditional, output_scale, probs, logBF, BF01, remove_spike_0, ...))
+    return(.summary.RoBMA.ss(object, type, conditional, output_scale, probs, logBF, BF01, remove_spike_0, standardized_coefficients, ...))
   }
 
   if(substr(type,1,1) == "e"){
@@ -187,6 +199,12 @@ summary.RoBMA       <- function(object, type = "ensemble", conditional = FALSE,
       }
 
       if(!is.null(object$RoBMA[["posteriors_predictors"]])){
+
+        # Transform samples if raw coefficients are requested
+        if (!standardized_coefficients && is.RoBMA.reg(object)){
+          object$RoBMA[["posteriors_predictors"]] <- .unstandardize_posterior_samples(object$RoBMA[["posteriors_predictors"]], attr(object$data$predictors, "variables_info"))
+        }
+
         # obtain estimates tables
         output$estimates_predictors <- BayesTools::ensemble_estimates_table(
           samples        = object$RoBMA[["posteriors_predictors"]],
@@ -212,6 +230,12 @@ summary.RoBMA       <- function(object, type = "ensemble", conditional = FALSE,
           attr(estimates_predictors_conditional, "footnotes") <- .scale_note(object$add_info[["prior_scale"]], output_scale)
           attr(estimates_predictors_conditional, "warnings")  <- .collect_errors_and_warnings(object)
         }else{
+
+          # Transform samples if raw coefficients are requested
+          if (!standardized_coefficients && is.RoBMA.reg(object)){
+            object$RoBMA[["posteriors_predictors_conditional"]] <- .unstandardize_posterior_samples(object$RoBMA[["posteriors_predictors_conditional"]], attr(object$data$predictors, "variables_info"))
+          }
+
           estimates_predictors_conditional <- BayesTools::ensemble_estimates_table(
             samples    = object$RoBMA[["posteriors_predictors_conditional"]],
             parameters = names(object$RoBMA[["posteriors_predictors_conditional"]]),
@@ -373,7 +397,7 @@ summary.RoBMA       <- function(object, type = "ensemble", conditional = FALSE,
   }
 }
 
-.summary.RoBMA.ss    <- function(object, type, conditional, output_scale, probs, logBF, BF01, remove_spike_0, ...){
+.summary.RoBMA.ss    <- function(object, type, conditional, output_scale, probs, logBF, BF01, remove_spike_0, standardized_coefficients, ...){
 
   if(substr(type,1,1) == "e"){
 
@@ -446,6 +470,12 @@ summary.RoBMA       <- function(object, type = "ensemble", conditional = FALSE,
       }
 
       if(!is.null(object$RoBMA[["posteriors_predictors"]])){
+
+        # Transform samples if raw coefficients are requested
+        if (!standardized_coefficients && is.RoBMA.reg(object)){
+          object$RoBMA[["posteriors_predictors"]] <- .unstandardize_posterior_samples(object$RoBMA[["posteriors_predictors"]], attr(object$data$predictors, "variables_info"))
+        }
+
         # obtain estimates tables
         output$estimates_predictors <- BayesTools::ensemble_estimates_table(
           samples        = object$RoBMA[["posteriors_predictors"]],
@@ -470,6 +500,12 @@ summary.RoBMA       <- function(object, type = "ensemble", conditional = FALSE,
           attr(estimates_predictors_conditional, "footnotes") <- .scale_note(object$add_info[["prior_scale"]], output_scale)
           attr(estimates_predictors_conditional, "warnings")  <- .collect_errors_and_warnings(object)
         }else{
+
+          # Transform samples if raw coefficients are requested
+          if (!standardized_coefficients && is.RoBMA.reg(object)){
+            object$RoBMA[["posteriors_predictors_conditional"]] <- .unstandardize_posterior_samples(object$RoBMA[["posteriors_predictors_conditional"]], attr(object$data$predictors, "variables_info"))
+          }
+
           estimates_predictors_conditional <- BayesTools::ensemble_estimates_table(
             samples    = object$RoBMA[["posteriors_predictors_conditional"]],
             parameters = names(object$RoBMA[["posteriors_predictors_conditional"]]),